backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.49",
    "fastapi>=0.135.1",
    "logfire>=3.0.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-ai[bedrock]>=0.4.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "uvicorn[standard]>=0.40.0",
]
